WIP: Attempt to centralize / generalize component configuration
OpenChakra originally has hard coded (chakraui related) symbols defined all over multiple files,
which seems redundant, so I took a chance to make it centralized.

Now componentDefs.tsx is the singe source of truth and for this trial some of the
chakraui components have already been added/configured this way.

componentDefs and componentDefDefaults contain all configurations of the components that
are to be displayed in the Sidebar and are draggable onto the Editor. The following variables
contain derived values (arrays and objects) that are used by other OPpenChakra modules:

- componentNames: name of all components listed in componentDefs
- rootComponentNames: name of root components, which may have children
- childComponentNames: name of child components, which have a parent
- menuItems: has the same structure as the original menuItems
- previewComponents: used in place of the switch in ComponentPreview.tsx
- rootDraggables: used in useDropComponent.tsx instead of rootComponents provided by editor.ts
- componentModelBuilders: used in useDropComponent.tsx instead of builders from builders.ts
- inspectorComponents: used in Panels.tsx instead of the quasi switch

With this centralized setup it should become easier to add new custom components to OpenChakra
or even replace the chakraui components with components of another toolkit (although the inspectors
are only aware of chakraui properties, however it follows styled-system more or less, so it
 should be feasible to use it for eg. NativeBase components).