Make style panels configurable via StylePanelsDef
The current set of style panels/inspectors that are specific for
chakraui are now configurable in terms of the actual style properties
to set. It is also possible to specify each style prop as
enabled/disabled in a given panel, or other configuration options maybe
passed to thse panels (see SpacingPanel)

A default StylePanelsDef is defined in

componentDefDefaults.stylePanelDef

and each component may configure an alternative StylePanelsDef
via stylePanelsDef or override the default one using
stylePanelsOverride.

Did some refactoring so now we have SpacingPanel, which displays two
PaddingPanels for maring and paddings. Also added a standalone
component for setting bg colors using BackgroundColorPanel.

With all this in effect it is now possible to reuse exising style
inspectors with alternative style props (eg. "bg" instead of
"backgroundColor") or to add new style inspectors, which will be picked
up by StylesPanel and passed along to each pannel in the panelDef prop.