Split core page builder functionality and UI kit configuration
The chakraui directory now only contains components that are related
to the handling of chakraui UI elements by OpenChakra. The connection
between the core builder UI and the chakraui elements it works with
is configured using a ComponentDefinitions instance
(a refactoring of the functionalities of  componentDefs.tsx).

There's a new ComponentDefinitionsProvider, which needs to be
added to the app and this needs to be configured with an instance
of ComponentDefinitions, which in turn needs to be created
using a ComponentDefs and ComponentDefDefault. These are now
provided by chakraui/componentDefs.tsx and this is the point where
a new ComponentDefinitions with a new ComponentDefs and
ComponentDefDefault can be provided for any other UI toolkit, eg.
NativeBase.

useComponentDefs() hook is used in the core builder components to
access the current UI kit component definitions and configurations.