More generalization and configuration options
OchoConfig/Ocho now handles onboarding templates and custom code
generation as well. For templates to be configurable app and container
model configuration is configured dynamicallt via initComponents and
initApp instead of intializing app and components global variables,
which cause too tight dependency between modules.

chakraui/componentDefs.tsx is an example of the current state of
configuraibilit, which result in the same functionality as
OpenChakra app.