Add code generation related props to the ComponentConfig
The "codeGenerationProps" property contains additional properties for the
components that are used during the source code generation.

The "codeGenerationOnly" prop shows whether the component is only used for
code generation. These components neither will be showed on the UI as droppable
components nor will be editable. These components will only be used to build
component previews that are not editable by the users.